+++ haiku_rag_slim/haiku/rag/graph/common/__init__.py
@@ -1,5 +1,5 @@
 """Common utilities for graph implementations."""
 
-from haiku.rag.graph.common.utils import get_model
+from haiku.rag.utils import get_model
 
 __all__ = ["get_model"]